# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "myCommerce",
    "description": "myCommerce API REST is a professional-grade application programming interface (API) designed specifically for an online commerce platform. It offers a comprehensive set of features and functionalities to facilitate seamless communication and data management between client applications and the commerce platform. [myCommerce repository](https://github.com/Danifantonrxd/myCommerce)",
    "contact": {
      "email": "danielrxd0912@gmail.com"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Add user",
        "description": "Endpoint to create an user in the database",
        "requestBody": {
          "required": true,
          "description": "Create user object",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "firstName",
                  "lastName"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@email.com",
                    "description": "must be a valid email"
                  },
                  "password": {
                    "type": "string",
                    "example": "myPassword",
                    "description": "must contain at least 8 characters"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "Emilia",
                    "description": "Min: 2, Max: 50 Characters"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Clarke",
                    "description": "Min: 2, Max: 50 Characters"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+11254789454"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successful operation, return the user missing the password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 5
                    },
                    "email": {
                      "type": "string",
                      "example": "user@email.com"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Emilia"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Clarke"
                    },
                    "role": {
                      "type": "string",
                      "example": "customer",
                      "description": "Used to grant access"
                    },
                    "phone": {
                      "type": "string",
                      "example": "+11254789454"
                    },
                    "recoveryToken": {
                      "type": "string",
                      "description": "this field is used to recover user password"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2023-09-25 22:33:01.621+00"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2023-09-25 22:33:01.621+00"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid product object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "message": {
                      "type": "string",
                      "example": "something is missing"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict with data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "example": "SequelizeUniqueConstraintError"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "email must be unique"
                          },
                          "type": {
                            "type": "string",
                            "example": "unique violation"
                          },
                          "...": {
                            "type": "string",
                            "example": "and so much properties"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "get all users",
        "description": "* Must be logged and have an admin user",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "jwt-token signed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": "15"
                      },
                      "email": {
                        "type": "string",
                        "example": "user@email.com"
                      },
                      "firstName": {
                        "type": "string",
                        "example": "Emilia"
                      },
                      "lastName": {
                        "type": "string",
                        "example": "Clarke"
                      },
                      "role": {
                        "type": "string",
                        "example": "customer",
                        "description": "Used to grant access"
                      },
                      "phone": {
                        "type": "string",
                        "example": "+11254789454"
                      },
                      "recoveryToken": {
                        "type": "string",
                        "description": "this field is used to recover user password"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2023-09-25 22:33:01.621+00"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2023-09-25 22:33:01.621+00"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "your role is not allow"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "get user by id",
        "description": "* Must be logged",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user_id you want to fetch. ",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 16
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "jwt-token signed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": "15"
                    },
                    "email": {
                      "type": "string",
                      "example": "user@email.com"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Emilia"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Clarke"
                    },
                    "role": {
                      "type": "string",
                      "example": "customer",
                      "description": "Used to grant access"
                    },
                    "phone": {
                      "type": "string",
                      "example": "+11254789454"
                    },
                    "recoveryToken": {
                      "type": "string",
                      "description": "this field is used to recover user password"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2023-09-25 22:33:01.621+00"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2023-09-25 22:33:01.621+00"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "your role is not allow"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not found"
                    },
                    "message": {
                      "type": "string",
                      "example": "user not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "update user by id",
        "description": "* Must be logged as the user you want to update or as an admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user_id you want to update. ",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 16
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "jwt-token signed"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Properties are optional",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@email.com",
                    "description": "must be a valid email"
                  },
                  "password": {
                    "type": "string",
                    "example": "myPassword",
                    "description": "must contain at least 8 characters"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "Emilia",
                    "description": "Min: 2, Max: 50 Characters"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Clarke",
                    "description": "Min: 2, Max: 50 Characters"
                  },
                  "role": {
                    "type": "string",
                    "example": "admin"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+11254789454"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation, user updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": "15"
                    },
                    "email": {
                      "type": "string",
                      "example": "user@email.com"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Emilia"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Clarke"
                    },
                    "role": {
                      "type": "string",
                      "example": "customer",
                      "description": "Used to grant access"
                    },
                    "phone": {
                      "type": "string",
                      "example": "+11254789454"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2023-09-25 22:33:01.621+00"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2023-09-25 22:33:01.621+00"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid user object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "message": {
                      "type": "string",
                      "example": "something is missing"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not found"
                    },
                    "message": {
                      "type": "string",
                      "example": "user not found"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict with data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "example": "SequelizeUniqueConstraintError"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "email must be unique"
                          },
                          "type": {
                            "type": "string",
                            "example": "unique violation"
                          },
                          "...": {
                            "type": "string",
                            "example": "and so much properties"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "delete user by id",
        "description": "* Must be logged as the user you want to delete or as an admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user_id you want to delete. ",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 16
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "jwt-token signed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation, user delete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": "15",
                      "description": "id of user that was deleted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "your role is not allow"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not found"
                    },
                    "message": {
                      "type": "string",
                      "example": "user not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}